Overwrote resupply implementation, command tweaks
Because func_regenerate also replaces their picked-up weapon with their
loadout's, we have to reimplement the resupply locker logic and account
for the custom weapon setup so it doesn't do all of that.

Also, enabling autogrant of jumpers grants you one instead of waiting
for a resupply (which may not happen).